Use float instead of double

+++ src/Shipwreck.Phash.CrossCorrelation/CrossCorrelation.cs
@@ -4,7 +4,7 @@ namespace Shipwreck.Phash
 {
     public static partial class CrossCorrelation
     {
-        public static double GetCrossCorrelation(byte[] coefficients1, byte[] coefficients2)
+        public static float GetCrossCorrelation(byte[] coefficients1, byte[] coefficients2)
             => GetCrossCorrelationCore(coefficients1, coefficients2, Math.Min(coefficients1.Length, coefficients2.Length));
 
         public static int GetHammingDistance(long x, long y)
